Create DeviceNET_testing.ino
I sniffed a PLC trainer that I had access to in order to figure out the handshake needed to control an Allen Bradley CompactBlock LDX (1790D-T8BV8B).  The notes are what I was able to figure out about the DeviceNET protocol available from the Internet at the time.  It is most certainly object oriented.